ingested more textures
there are a lot of them that need to be brought in line with the naming convention for PBR to work, but most of the files are at least in the correct places for Godot and Qodot to use.

lots of _normal_ogl that need to just be _normal, dashes that need to be underscores.

There are also probably some misplaced albedos. Late in the organizing process I decided to have a misc category.